fix: 🐛 Small fix when writing floats

+++ packages/exporter/src/utils/Float.ts
@@ -4,7 +4,7 @@ export class Float {
     }
 
     encodeCBOR(encoder: any) {
-        return encoder._pushFloat(this.value)
+        return encoder._pushFloat(Math.fround(this.value))
     }
 }
